Refactored away CompanyIdVm. Moved typescript interfaces to separate files.
