+++ web/src/main/kotlin/no/ssb/kostra/web/viewmodel/KostraFormVm.kt
@@ -3,7 +3,6 @@ package no.ssb.kostra.web.viewmodel
 import io.micronaut.core.annotation.Introspected
 import no.ssb.kostra.web.validation.GyldigSkjema
 import no.ssb.kostra.web.validation.GyldigSkjemaType
-import javax.validation.Valid
 import javax.validation.constraints.Min
 import javax.validation.constraints.NotBlank
 import javax.validation.constraints.Pattern
@@ -23,11 +22,11 @@ data class KostraFormVm(
 
     val navn: String? = null,
 
-    @field:Valid
-    val orgnrForetak: CompanyIdVm? = null,
+    @field:Pattern(regexp = "[8|9]\\d{8}", message = "Må starte med [8|9] etterfulgt av 8 siffer")
+    val orgnrForetak: String? = null,
 
-    @field:Valid
-    val orgnrVirksomhet: Collection<CompanyIdVm>? = null,
+    /** When Micronaut 4, add validation to each entry like orgnrForetak */
+    val orgnrVirksomhet: Collection<String>? = null,
 
     @field:NotBlank(message = "Filvedlegg mangler")
     val base64EncodedContent: String? = null
