+++ web/src/main/vite-project/src/kostratypes/errorDetailsVm.ts
@@ -1,0 +1,9 @@
+import {KostraErrorCode} from "./kostraErrorCode";
+
+export interface ErrorDetailsVm {
+    readonly journalnummer: NonNullable<string>
+    readonly saksbehandler: NonNullable<string>
+    readonly kontrollnummer: NonNullable<string>
+    readonly kontrolltekst: NonNullable<string>
+    readonly feilkode: NonNullable<KostraErrorCode>
+}
