+++ web/src/main/vite-project/src/kostratypes/kostraErrorCode.ts
@@ -1,0 +1,7 @@
+export enum KostraErrorCode {
+    NO_ERROR = 0,
+    NORMAL_ERROR,
+    CRITICAL_ERROR,
+    SYSTEM_ERROR,
+    PARAMETER_ERROR
+}
