+++ web/src/main/vite-project/src/kostratypes/nullable.ts
@@ -1,0 +1,1 @@
+export type Nullable<T> = T | null
